Make Handlebars::Context.current thread safe
Handlebars::Template#call was mutating global state. When called in
multiple threads you ran the risk of overwriting the current context for
other threads while the @fn.call was executing. By moving to a thread
variable we ensure that this value remains with the thread which is
setting the value.

Also in this PR are some dependency updates including requiring a modern
ruby. All three ruby versions previously tested are no longer
maintained.